+++ threedi_schematisation_editor/forms/custom_forms.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2025 by Lutra Consulting
+# Copyright (C) 2025 by Nelen & Schuurmans
 import sys
 from collections import defaultdict
 from enum import Enum