+++ threedi_schematisation_editor/processing/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2025 by Lutra Consulting
+# Copyright (C) 2025 by Nelen & Schuurmans
 import os
 
 from qgis.core import QgsProcessingProvider
@@ -42,11 +42,13 @@ class ThreediSchematisationEditorProcessingProvider(QgsProcessingProvider):
         return "threedi_schematisation_editor"
 
     def name(self):
-        return "3Di Schematisation Editor"
+        return "Rana Schematisation Editor"
 
     def icon(self):
         icon_path = os.path.join(
-            os.path.dirname(os.path.dirname(__file__)), "icons", "icon.png"
+            os.path.dirname(os.path.dirname(__file__)),
+            "icons",
+            "rana_blue_on_transparent.svg",
         )
         return QIcon(icon_path)