+++ threedi_schematisation_editor/processing/algorithms_1d.py
@@ -1,4 +1,4 @@
-# Copyright (C) 2025 by Lutra Consulting
+# Copyright (C) 2025 by Nelen & Schuurmans
 from collections import defaultdict
 
 from qgis.core import (
@@ -48,10 +48,10 @@ class BottomLevelCalculator(QgsProcessingAlgorithm):
         <p>For each connection node manhole, the algorithm determines which sides of which pipes (or culverts) are connected to it, and what the invert level is at that side. It than takes the lowest of these invert levels as bottom level for the manhole.</p>
         <h3>Parameters</h3>
         <h4>Connection node layer</h4>
-        <p>Connection node layer that is added to the project with the 3Di Schematisation Editor.</p>
+        <p>Connection node layer that is added to the project with the Rana Schematisation Editor.</p>
         <p>If "Selected connection nodes only" is checked, only the selected manholes will be used in the algorithm.</p>
         <h4>Pipe layer</h4>
-        <p>Pipe or Culvert layer that is added to the project with the 3Di Schematisation Editor.</p>
+        <p>Pipe or Culvert layer that is added to the project with the Rana Schematisation Editor.</p>
         <p>If "Selected pipes only" is checked, only the selected pipes will be used in the algorithm.</p>
         <h4>Overwrite existing bottom levels</h4>
         <p>If checked, bottom levels will be recalculated for manholes that already have a bottom level filled in.</p>